chore: adjust some logic
