+++ moon/apps/web/components/Issues/IssueNewPage.tsx
@@ -18,15 +18,13 @@ import { useScope } from '@/contexts/scope'
 import { usePostIssueSubmit } from '@/hooks/issues/usePostIssueSubmit'
 import { useGetCurrentUser } from '@/hooks/useGetCurrentUser'
 import { useGetOrganizationMember } from '@/hooks/useGetOrganizationMember'
-import { useSyncedMembers } from '@/hooks/useSyncedMembers'
 import { useUploadHelpers } from '@/hooks/useUploadHelpers'
 import { apiErrorToast } from '@/utils/apiErrorToast'
 import { trimHtml } from '@/utils/trimHtml'
 
 import { MemberAvatar } from '../MemberAvatar'
-import { tags } from './utils/consts'
-import { extractTextArray } from './utils/extractText'
 import { pickWithReflect } from './utils/pickWithReflectDeep'
+import { splitFun, useAvatars, useLabelMap, useLabels, useMemberMap } from './utils/sideEffect'
 
 // import { OrganizationMember } from '@gitmono/types/generated'
 
@@ -54,39 +52,10 @@ export default function IssueNewPage() {
     }
   }, [user])
 
-  const splitFun = (el: React.ReactNode): string[] => {
-    return extractTextArray(el)
-      .flatMap((name) => name.split(',').map((n) => n.trim()))
-      .filter((n) => n.length > 0)
-  }
-
-  const { members } = useSyncedMembers()
-
-  const avatars: ItemInput[] = useMemo(
-    () =>
-      members?.map((i) => ({
-        groupId: 'end',
-        text: i.user.display_name,
-        leadingVisual: () => <MemberAvatar size='sm' member={i} />
-      })) || [],
-    [members]
-  )
+  const avatars = useAvatars()
   // const [avatarTems, setAvatarItems] = useState<ItemInput[]>(avatars)
 
-  const labels: ItemInput[] = useMemo(
-    () =>
-      tags.map((i) => ({
-        text: i.description,
-        leadingVisual: () => (
-          <div
-            className='h-[14px] w-[14px] rounded-full border'
-            //eslint-disable-next-line react/forbid-dom-props
-            style={{ backgroundColor: i.color, borderColor: i.color }}
-          />
-        )
-      })),
-    []
-  )
+  const labels = useLabels()
 
   const [isReactionPickerOpen, setIsReactionPickerOpen] = useState(false)
   const set_to_loading = (index: number) => {
@@ -137,23 +106,9 @@ export default function IssueNewPage() {
     upload: editorRef.current?.uploadAndAppendAttachments
   })
 
-  const memberMap = useMemo(() => {
-    const map = new Map()
+  const memberMap = useMemberMap()
 
-    members?.forEach((i) => {
-      map.set(i.user.display_name, i)
-    })
-    return map
-  }, [members])
-
-  const labelMap = useMemo(() => {
-    const map = new Map()
-
-    tags.map((i) => {
-      map.set(i.description, i)
-    })
-    return map
-  }, [])
+  const labelMap = useLabelMap()
 
   // const createGroup = (selected: ItemInput[]) => {
   //   setAvatarItems((prev) => {
